backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Werkzeug
python-dotenv
faiss-cpu
torch
transformers
dotenv
pymupdf 
python-docx 
pytesseract 
requests 
beautifulsoup4 
youtube-transcript-api 
pillow
flask-migrate
openai
flask-cors